+++ workers/unity/Packages/io.improbable.gdk.core/Worker/WorkerConnector.cs
@@ -265,35 +265,9 @@ namespace Improbable.Gdk.Core
 
         private static ushort GetPlayerConnectionPort()
         {
-            var companyName = Application.companyName;
-            var productName = Application.productName;
-
-            string logPath;
-
-            switch (Application.platform)
-            {
-                case RuntimePlatform.WindowsPlayer:
-                    logPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "AppData", "LocalLow", companyName, productName,
-                        "Player.log");
-                    break;
-                case RuntimePlatform.OSXPlayer:
-                    // On MacOS it always goes in the Unity folder regardless of what companyName and productName are set to.
-                    logPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), "Library", "Logs", "Unity", "Player.log");
-                    break;
-                case RuntimePlatform.LinuxPlayer:
-                    logPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.UserProfile), ".config", "unity3d", companyName, productName, "Player.log");
-                    break;
-                default:
-                    throw new InvalidOperationException($"Cannot find log file on platform: {Application.platform}");
-            }
-
-            CommandLineArgs.FromCommandLine().TryGetCommandLineValue("logfile", ref logPath);
-
-            logPath = Path.GetFullPath(logPath);
-
             // We need to open the File as ReadWrite since this process _already_ has it open as ReadWrite.
             // Attempting to open it as Read only results in IO exceptions due to permissions. Go figure.
-            using (var stream = new FileStream(logPath, FileMode.Open, FileAccess.ReadWrite, FileShare.ReadWrite))
+            using (var stream = new FileStream(Application.consoleLogPath, FileMode.Open, FileAccess.ReadWrite, FileShare.ReadWrite))
             using (var readStream = new StreamReader(stream))
             {
                 var logContents = readStream.ReadToEnd();