+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Added
 
-- Added support for Unity build callbacks such as `IPostprocessBuildWithReport`.
-  - During a build you will have access to the `WorkerBuilder.CurrentContext` field which contains all SpatialOS related build information.
+- Added support for Unity build callbacks such as `IPostprocessBuildWithReport`. [#1228](https://github.com/spatialos/gdk-for-unity/pull/1228)
+  - During a build you can now access the `WorkerBuilder.CurrentContext` field which contains all SpatialOS related build information.
 
 ### Fixed